Refactor: adjusting location of ROUTE TASK validations

+++ src/main/java/com/electricitybill/service/impl/TarefaRotaServiceImpl.java
@@ -3,10 +3,12 @@ package com.electricitybill.service.impl;
 import com.electricitybill.entity.TarefaRotaEntity;
 import com.electricitybill.repository.TarefaRotaRepository;
 import com.electricitybill.service.TarefaRotaService;
+import jakarta.persistence.NoResultException;
 import lombok.AllArgsConstructor;
 
 import java.sql.SQLException;
 import java.util.List;
+import java.util.Objects;
 
 @AllArgsConstructor
 public class TarefaRotaServiceImpl implements TarefaRotaService {
@@ -15,11 +17,17 @@ public class TarefaRotaServiceImpl implements TarefaRotaService {
 
     @Override
     public List<TarefaRotaEntity> findAll() {
+        if (tarefaRotaRepository.findAll().isEmpty()) {
+            throw new NoResultException("Search not found!");
+        }
         return tarefaRotaRepository.findAll();
     }
 
     @Override
     public TarefaRotaEntity findById(int id) {
+        if (Objects.isNull(tarefaRotaRepository.findById(id))) {
+            throw new NoResultException("Individual search not found!");
+        }
         return tarefaRotaRepository.findById(id);
     }
 
@@ -30,11 +38,17 @@ public class TarefaRotaServiceImpl implements TarefaRotaService {
 
     @Override
     public boolean update(int id, TarefaRotaEntity entity) throws SQLException {
+        if (Objects.isNull(tarefaRotaRepository.findById(id))) {
+            throw new NoResultException("Route task not found!");
+        }
         return tarefaRotaRepository.update(id, entity);
     }
 
     @Override
     public boolean delete(int id) throws SQLException {
+        if (Objects.isNull(tarefaRotaRepository.findById(id))) {
+            throw new NoResultException("Route task not found!");
+        }
         return tarefaRotaRepository.delete(id);
     }
 }
